refactor: rename interpolation hook
