+++ frontend/src/features/power_unit/hooks/usePropellerInterpolation.ts
